# backend-v3-grpc-roles/docs/swagger.yaml
basePath: /
definitions:
  dto.AuthResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserData'
      message:
        example: User registered successfully
        type: string
      success:
        example: true
        type: boolean
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  dto.CreateLinkRequest:
    properties:
      description:
        example: Search engine
        type: string
      title:
        example: Google
        type: string
      url:
        example: https://google.com
        type: string
    required:
    - title
    - url
    type: object
  dto.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: johndoe
        type: string
    required:
    - password
    - username
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      username:
        example: johndoe
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.Response:
    properties:
      count:
        example: 10
        type: integer
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.UpdateLinkRequest:
    properties:
      description:
        example: Updated description
        type: string
      title:
        example: Updated Title
        type: string
      url:
        example: https://example.com
        type: string
    type: object
  dto.UserData:
    properties:
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      username:
        example: johndoe
        type: string
    type: object
host: localhost:8082
info:
  contact:
    email: support@linkmanager.com
    name: API Support
  description: REST API для управления ссылками с JWT аутентификацией (Onion Architecture)
    на Go + SQLite
  title: Link Manager API with Auth
  version: "2.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Вход пользователя
      tags:
      - Auth
  /api/auth/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Получить профиль текущего пользователя
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Регистрация нового пользователя
      tags:
      - Auth
  /api/links:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Получить все ссылки пользователя
      tags:
      - Links
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные ссылки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Создать новую ссылку
      tags:
      - Links
  /api/links/{id}:
    delete:
      parameters:
      - description: Link ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Удалить ссылку
      tags:
      - Links
    get:
      parameters:
      - description: Link ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Получить ссылку по ID
      tags:
      - Links
    put:
      consumes:
      - application/json
      parameters:
      - description: Link ID
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Обновить ссылку
      tags:
      - Links
  /api/links/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Получить статистику пользователя
      tags:
      - Links
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"