# backend-v3-grpc-roles/internal/presentation/dto/auth_dto.go
package dto

// RegisterRequest представляет запрос на регистрацию
type RegisterRequest struct {
	Username string `json:"username" binding:"required" example:"johndoe"`
	Email    string `json:"email" binding:"required,email" example:"john@example.com"`
	Password string `json:"password" binding:"required,min=6" example:"password123"`
}

// LoginRequest представляет запрос на вход
type LoginRequest struct {
	Username string `json:"username" binding:"required" example:"johndoe"`
	Password string `json:"password" binding:"required" example:"password123"`
}

// AuthResponse представляет ответ при аутентификации
type AuthResponse struct {
	Success bool        `json:"success" example:"true"`
	Data    *UserData   `json:"data,omitempty"`
	Token   string      `json:"token,omitempty" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`
	Message string      `json:"message,omitempty" example:"User registered successfully"`
}

// UserData представляет данные пользователя
type UserData struct {
	ID       uint   `json:"id" example:"1"`
	Username string `json:"username" example:"johndoe"`
	Email    string `json:"email" example:"john@example.com"`
}